Integrated IPC cache and XID push notifications, issue #546 (#344)
* checkpoint

* checkout

* checkpoint

* checkpoint

* checkpoint

* code complete

* checkpoint (#340)

* fix: test improvement (#341)

* fix: test improvement

* tar the unit test and integration test logs

* fix: missing .tar.gz extension

* Fixes SPR-618 (#342)

* Jeff/fix nightly (#343)

* Fixes the slack notification

* grr

* fix more

* dead code

* sonarqube

* more sonarqube

* checkpoint

* check for table state

* removed old file

* Addressed comments

* Added coordinator keep alive

* Fixed the PropertiesRunner flag

* optimized the lock scope

---------

+++ python/coordinator/runit.py
@@ -89,9 +89,14 @@ def run_tests(factory: ComponentFactory) -> None:
     xid_mgr_daemon.start()
     assert xid_mgr_daemon.is_running()
 
+    pg_xid_subscriber_daemon = factory.create_pg_xid_subscriber_daemon()
+    test(pg_xid_subscriber_daemon)
+    assert not sys_tbl_mgr_daemon.is_running()
+
     sys_tbl_mgr_daemon = factory.create_sys_tbl_mgr_daemon()
     test(sys_tbl_mgr_daemon)
     assert not sys_tbl_mgr_daemon.is_running()
+
 
     postgres = factory.create_postgres()
     test(postgres)
@@ -151,6 +156,9 @@ if __name__ == "__main__":
         sys_tbl_mgr_daemon = factory.create_sys_tbl_mgr_daemon()
         if not sys_tbl_mgr_daemon.shutdown() and not sys_tbl_mgr_daemon.kill():
             raise ValueError("Failed to stop sys_tbl_mgr_daemon")
+        pg_xid_subscriber_daemon = factory.create_pg_xid_subscriber_daemon()
+        if not pg_xid_subscriber_daemon.shutdown() and not pg_xid_subscriber_daemon.kill():
+            raise ValueError("Failed to stop pg_xid_subscriber_daemon")
         ddl_daemon = factory.create_ddl_daemon('test_user', 'test_password')
         if not ddl_daemon.shutdown() and not ddl_daemon.kill():
             raise ValueError("Failed to stop ddl_daemon")
